refactor: use SafeAreaView within headers
