+++ src/components/Layout/Header/AvatarHeader/index.tsx
@@ -1,14 +1,18 @@
 import React from "react";
+import { SafeAreaView } from "react-native";
 import { TouchableOpacity } from "react-native-gesture-handler";
 
 import useNavigate from "~/hooks/useNavigate";
 import { PROFILE_ROUTE } from "~/router/routes";
-import { HeaderContainer } from "~/components/Layout/Header/styles";
+import  HeaderContainer from "~/components/Layout/Header/HeaderContainer";
 
 import { HeaderUserAvatar } from "./styles";
 import { AvatarHeaderProps } from "./types";
 
-const AvatarHeader: React.FC<AvatarHeaderProps> = ({ children, ...rest }) => {
+const AvatarHeader: React.FC<AvatarHeaderProps> = ({
+  children,
+  ...rest
+}) => {
   const navigate = useNavigate();
 
   return (
